fixed movement/interpolation from client side

+++ Client/Assets/Scripts/Client/Components.cs
@@ -18,8 +18,8 @@ namespace Client
     
     public struct Unit : IComponentData
     {
-        public uint unitId;
-        public uint player;
+        public ushort unitId;
+        public byte player;
         public bool isActivePlayer;
         public bool isSelected;
     }
